+++ src/components/Viewer/Viewer.tsx
@@ -126,6 +126,8 @@ interface ViewerProps {
   variableSplitPos?: number;
   onMapRef?: (map: OlMap | null) => void;
   importUserPlacesFromText?: (text: string) => void;
+  setZoomLevel?: (zoomLevel: number | undefined) => void;
+  zoomBox?: MapElement;
 }
 
 export default function Viewer({
@@ -158,6 +160,8 @@ export default function Viewer({
   imageSmoothing,
   variableSplitPos,
   onMapRef,
+  zoomBox,
+  setZoomLevel,
 }: ViewerProps) {
   theme = useTheme();
 
@@ -279,6 +283,14 @@ export default function Viewer({
     }
   };
 
+  const handleMapZoom = (event: OlMapBrowserEvent<UIEvent>) => {
+    const zoomLevel = event.target.getZoom();
+
+    if (setZoomLevel) {
+      setZoomLevel(zoomLevel);
+    }
+  };
+
   const handleDrawEnd = (event: DrawEvent) => {
     if (map !== null && addDrawnUserPlace && mapInteraction !== "Select") {
       const feature = event.feature;
@@ -356,6 +368,7 @@ export default function Viewer({
       <Map
         id={mapId}
         onClick={(event) => handleMapClick(event)}
+        onZoom={(event) => handleMapZoom(event)}
         onMapRef={handleMapRef}
         mapObjects={MAP_OBJECTS}
         isStale={true}
@@ -426,6 +439,7 @@ export default function Viewer({
         {mapPointInfoBox}
         {mapControlActions}
         {mapSplitter}
+        {zoomBox}
         <ScaleLine bar={false} />
       </Map>
     </ErrorBoundary>