+++ src/states/controlState.ts
@@ -136,6 +136,8 @@ export interface ControlState {
   exportFileName: string;
   themeMode: ThemeMode;
   exportResolution: ExportResolution;
+  showZoomBox: boolean;
+  zoomLevel: number | undefined;
 }
 
 export function newControlState(): ControlState {
@@ -217,6 +219,8 @@ export function newControlState(): ControlState {
     exportFileName: "export",
     themeMode: getInitialThemeMode(),
     exportResolution: 300,
+    showZoomBox: branding.showZoomBox || false,
+    zoomLevel: undefined,
   };
   return loadUserSettings(state);
 }