+++ CHANGES.md
@@ -27,7 +27,7 @@
 ### Fixes
 
 *  Applied workaround for a bug in `html-to-image` libary that causes an 
-   issuesfor the export of screenshorts (charts and map) in the Firefox 
+   issue for the export of screenshorts (charts and map) in the Firefox 
    browser.
    
 ### New Features 
@@ -42,6 +42,13 @@
   Window. If this feature is configured, an `About` window can be opened
   with a button in the header and it will be shown initially while
   data is loading. (#508) 
+  
+* A zoom-level indicator was added to the map. This box displays the current 
+  zoom level of the map and the dataset resolution level used for displaying it. 
+  The visibility of this feature can be controlled in the settings. The initial visibility 
+  can be set in `config.json` (`"branding":{ "showZoomBox": true, ...`), 
+  the default is `false`. In addition, the total number of dataset levels has been 
+  added to the metadata in the Info panel. (#287)
 
 ### Other changes